fix(services): resolve type mismatches and borrow errors for clean build
- payroll_service: use i64 for total_amount accumulator and salary param;
  cast treasury SUM to ::BIGINT to fix NUMERIC type mismatch from COALESCE;
  clone payroll before move in execute_payroll to fix borrow-after-move
- merkle_tree: update MerkleTree::new, hash_leaf, and get_proof to use
  i64 amounts instead of i128
- zk_prover: update generate_commitment_hash, generate_payroll_proof,
  and generate_salary_commitment signatures to use i64

+++ src/services/utils/merkle_tree.rs
@@ -9,7 +9,7 @@ pub struct MerkleTree {
 }
 
 impl MerkleTree {
-    pub fn new(data: Vec<(i64, i128)>) -> Self {
+    pub fn new(data: Vec<(i64, i64)>) -> Self {
         let mut leaves = Vec::new();
         let mut leaf_map = HashMap::new();
         
@@ -30,7 +30,7 @@ impl MerkleTree {
         }
     }
 
-    fn hash_leaf(id: i64, amount: i128) -> String {
+    fn hash_leaf(id: i64, amount: i64) -> String {
         let mut hasher = Sha256::new();
         hasher.update(id.to_be_bytes());
         hasher.update(amount.to_be_bytes());
@@ -74,7 +74,7 @@ impl MerkleTree {
         self.root.clone()
     }
 
-    pub fn get_proof(&self, id: i64, amount: i128) -> Vec<String> {
+    pub fn get_proof(&self, id: i64, amount: i64) -> Vec<String> {
         let leaf = Self::hash_leaf(id, amount);
         let mut proof = Vec::new();
